Update to xemu v0.312b
This update is mainly cleaning up (well, "cleaning up") code and making it a little more bearable inside. There's a speed hit, but it's only like 5%, and it comes with a slight revamp in the benchmark command to preserve the current emulator state. There's also a memory usage increase and it's a few thousand bytes larger... oh well.